requirements.txt
speechrecognition
scipy
dtw-python
numpy
